// src/ontologies/xsd.d.ts
import { NamedNode } from '../models.js';
export declare var _ontologyResource: NamedNode;
export declare const xsd: {
    _ontologyResource: NamedNode;
    string: NamedNode;
    boolean: NamedNode;
    date: NamedNode;
    integer: NamedNode;
    time: NamedNode;
    duration: NamedNode;
    decimal: NamedNode;
    gYear: NamedNode;
    Bytes: NamedNode;
    long: NamedNode;
};