// src/ontologies/rdfs.d.ts
import { NamedNode } from '../models.js';
export declare var _ontologyResource: NamedNode;
export declare var subPropertyOf: NamedNode;
export declare const rdfs: {
    _ontologyResource: NamedNode;
    subPropertyOf: NamedNode;
    subClassOf: NamedNode;
    range: NamedNode;
    isDefinedBy: NamedNode;
    label: NamedNode;
    Literal: NamedNode;
    Datatype: NamedNode;
    Class: NamedNode;
    Resource: NamedNode;
};